+++ code/utils/index.ts
@@ -1,0 +1,5 @@
+export async function delay(duration: number) {
+    return new Promise((resolve, reject) => {
+        setTimeout(resolve, duration)
+    })
+}
